[UPDATE]:Improving error handling

+++ src/run.py
@@ -18,6 +18,11 @@ class GetPath:
 
 class Input:
     def __init__(self)->None:
+
+        if not os.path.exists(sys.argv[1]) and not os.path.exists(sys.argv[2]):
+            print("Error: The binary file or icon file don't exist, or cannot be reached")
+            sys.exit(2);
+
         self.bin_abs: Path = Path(sys.argv[1]).resolve();
         self.dsk_abs: Path = Path(sys.argv[2]).resolve();
         self.f_name: Path = Path(sys.argv[3]);
@@ -40,35 +45,24 @@ def make_symlink(src:Path, dst:Path)->None:
     os.symlink(src, dst);
 
 def remove_files() -> None:
-    if sys.argv[1] == "rm":
+    if sys.argv[1] == "rm" and os.path.exists(sys.argv[2]):
         gp = GetPath();
         tgt = TargetPath(gp);
 
         unlink(tgt.tgt_bin_path);
         unlink(tgt.tgt_dsk_path);
+    else:
+        print("\033[91mFailed: \033[0mThe shortcut don't exist");
+        sys.exit(3);
 
 def main() -> None:
-
-    if len(sys.argv) <= 2 or len(sys.argv) > 4:
-        print("\nusage:\n'desktop <executable file> <icon> <name>' to create a shortcut \n or \n'desktop rm <name>' to remove the shortcut \n");
-        return;
-
-    if sys.argv[1] == "rm" and len(sys.argv) == 3 :
-        remove_files();
-        print(f"Shortcut for {sys.argv[2]} Removed");
-        return;
-
-    elif sys.argv[1] == "rm" and len(sys.argv) != 3:
-        print("\nusage:\n'desktop <executable file> <icon> <name>' to create a shortcut \n or \n'desktop rm <name>' to remove the shortcut \n");
-        return;
 
     gp = GetPath();
     ui = Input();
     tgt = TargetPath(gp);
 
     make_symlink(ui.bin_abs, tgt.tgt_bin_path);
-    print(f"Symlink created: {tgt.tgt_bin_path} -> {os.readlink(tgt.tgt_bin_path)}");
-    print("Shortcut created\n");
+    print(f"\n\033[92mSymlink created: \033[0m{tgt.tgt_bin_path} \033[92m->\033[0m {os.readlink(tgt.tgt_bin_path)}\n");
 
     with open(tgt.tgt_dsk_path, "w") as f:
         f.write("[Desktop Entry]\n"
@@ -81,9 +75,17 @@ def main() -> None:
             f"Comment={ui.f_name}\n"
             f"Categories=Utility;\n");
         f.close();
-
+        
     st = os.stat(tgt.tgt_bin_path);
     os.chmod(tgt.tgt_bin_path, st.st_mode | stat.S_IEXEC | stat.S_IXGRP | stat.S_IXUSR);
+    print("\033[92mShortcut created\n");
 
 if __name__ == "__main__":
-    main();
+    match sys.argv[1:]:
+        case ["rm", file]:
+            remove_files();
+        case [file, icon, name]:
+            main();
+        case _:
+            print("\n\033[93musage:\033[0m\n'desktop <executable file> <icon> <name>' to create a shortcut \n or \n'desktop rm <name>' to remove the shortcut \n");
+            sys.exit(1);
